requirements.txt
qiskit>=2.0
qiskit-aer>=0.15
qiskit-ibm-runtime>=0.30
numpy
matplotlib
Pillow
opencv-python-headless
streamlit>=1.30
pylatexenc
pyyaml
